change PayPal logic
